fix click handler and put inputs in a form
The click handler (er, now submit handler) must return false to
prevent the propagation of the submit event beyond our handler.
The fact that "return false" appeared outside of the handler was
probably a typo.

I put the inputs in a form so that the enter key will cause the
submit button to be clicked, which is the usual preferred
behavior.