fix exception in tag formation
The sample code on which the saveTrail function is based sets the
bookmark.data("tags") to the existing tags of a bookmark when it
is fetched from delicious.  But because our bookmarks are build
from scratch, we were not setting "tags".  So dereferencing it
was causing an exception.  This was hard to catch because the
exception prevented the .submit handler from returning false,
which in turn caused the browser to do its default thing, which
is to redirect and post to the page specified by the action
property of the form.