integrate viewer and use redirect instead of (in)visible divs
The existing code used visible and hidden divs to handle the page
flow.  But this breaks the user's back button.  Also, integrating
the css and js used in the viewer turned out to be a bit
problematic.  Specifically, the styles were interfering, and I
was getting unexpected behavior from coverflow.